Create next layout
